requirements.txt
python-dotenv==1.1.0
streamlit==1.45.0
langchain==0.3.23
langchain-community==0.3.21
langchain-core==0.3.58
langchain-openai==0.3.16
langchain-text-splitters==0.3.8
faiss-cpu==1.11.0
boto3==1.38.8
